# Improved performance of PropagationBP