--- a/src/module/PropagationBP.cpp
+++ b/src/module/PropagationBP.cpp
@@ -6,7 +6,7 @@
 
 namespace crpropa {
 	void PropagationBP::tryStep(const Y &y, Y &out, Y &error, double h,
-			ParticleState &particle, double z, double m, double q) const {
+			ParticleState &particle, double z, double q, double m) const {
 		out = dY(y.x, y.u, h, z, q, m);  // 1 step with h
 
 		Y outHelp = dY(y.x, y.u, h/2, z, q, m);  // 2 steps with h/2
@@ -95,7 +95,7 @@ namespace crpropa {
 
 			// try performing step until the target error (tolerance) or the minimum/maximum step size has been reached
 			while (true) {
-				tryStep(yIn, yOut, yErr, step, current, z, m, q);
+				tryStep(yIn, yOut, yErr, step, current, z, q, m);
 				r = yErr.u.getR() / tolerance;  // ratio of absolute direction error and tolerance
 				if (r > 1) {  // large direction error relative to tolerance, try to decrease step size
 					if (step == minStep)  // already minimum step size