--- a/test/testPropagation.cpp
+++ b/test/testPropagation.cpp
@@ -3,6 +3,14 @@
 #include "crpropa/module/SimplePropagation.h"
 #include "crpropa/module/PropagationBP.h"
 #include "crpropa/module/PropagationCK.h"
+#include "crpropa/Grid.h"
+#include "crpropa/Units.h"
+#include "crpropa/Common.h"
+#include "crpropa/GridTools.h"
+#include "crpropa/magneticField/turbulentField/TurbulentField.h"
+#include "crpropa/magneticField/turbulentField/GridTurbulence.h"
+#include "crpropa/magneticField/turbulentField/PlaneWaveTurbulence.h"
+#include "crpropa/magneticField/turbulentField/SimpleGridTurbulence.h"
 
 #include "gtest/gtest.h"
 
@@ -481,6 +489,44 @@ TEST(testPropagationBP, gyration) {
 	EXPECT_DOUBLE_EQ(2 / 3., dirX * dirX + dirY * dirY);  // constant momentum in the perpendicular plane to background magnetic field field
 	EXPECT_DOUBLE_EQ(1 / 3., dirZ * dirZ);  // constant momentum parallel to the background magnetic field
 	EXPECT_DOUBLE_EQ(100 * step * step / 3., posZ * posZ);  // constant velocity parallel to the background magnetic field
+}
+
+
+// Test the that the optimization for same step sizes works
+TEST(testPropagationBP, sameStepOptimization) {
+
+	PropagationBP propa1(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), pc);
+	ParticleState p1;
+	p1.setId(nucleusId(1, 1));
+	p1.setEnergy(100 * EeV);
+	p1.setPosition(Vector3d(0, 0, 0));
+	p1.setDirection(Vector3d(1, 1, 1));
+	Candidate c1(p1);
+	c1.setNextStep(0);
+	// Nine new steps to have finally propagated the particle ten times
+	for (int i = 0; i < 9; i++){
+		propa1.process(&c1);
+	}
+
+	PropagationBP propa2(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), 1, pc, 1.1*pc);
+	ParticleState p2;
+	p2.setId(nucleusId(1, 1));
+	p2.setEnergy(100 * EeV);
+	p2.setPosition(Vector3d(0, 0, 0));
+	p2.setDirection(Vector3d(1, 1, 1));
+	Candidate c2(p2);
+	c1.setNextStep(0);
+	// Nine new steps to have finally propagated the particle ten times
+	for (int i = 0; i < 9; i++){
+		propa2.process(&c2);
+	}
+
+	EXPECT_DOUBLE_EQ(c1.current.getDirection().x, c2.current.getDirection().x);
+	EXPECT_DOUBLE_EQ(c1.current.getDirection().y, c2.current.getDirection().y);
+	EXPECT_DOUBLE_EQ(c1.current.getDirection().z, c2.current.getDirection().z);
+	EXPECT_DOUBLE_EQ(c1.current.getPosition().x, c2.current.getPosition().x);
+	EXPECT_DOUBLE_EQ(c1.current.getPosition().y, c2.current.getPosition().y);
+	EXPECT_DOUBLE_EQ(c1.current.getPosition().z, c2.current.getPosition().z);
 }