--- a/test/testPropagation.cpp
+++ b/test/testPropagation.cpp
@@ -485,12 +485,11 @@ TEST(testPropagationBP, gyration) {
 }
 
 
-// Test the that the optimization for same step sizes works
-TEST(testPropagationBP, sameStepOptimization) {
-
+// Test the that the optimization for fixed step sizes works
+TEST(testPropagationBP, fixedStepOptimization) {
 	// particle 1 with fixed step sizes
-	double fixeded_step = pc;
-	PropagationBP propa1(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), fixeded_step);
+	double fixed_step = pc;
+	PropagationBP propa1(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), fixed_step);
 	ParticleState p1;
 	p1.setId(nucleusId(1, 1));
 	p1.setEnergy(100 * EeV);
@@ -506,7 +505,7 @@ TEST(testPropagationBP, sameStepOptimization) {
 	// particle 2 with different min and max steps. The tolerance is chosen such that particle 2 will be
 	// propagated with the same step as particle 1, however not using the optimization for fixed step sizes
 	double tolerance = 1;
-	PropagationBP propa2(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), tolerance, fixeded_step, 1.1*fixeded_step);
+	PropagationBP propa2(new PlaneWaveTurbulence(TurbulenceSpectrum(gauss, pc, 100*pc), 10, 1), tolerance, fixed_step, 1.1*fixed_step);
 	ParticleState p2;
 	p2.setId(nucleusId(1, 1));
 	p2.setEnergy(100 * EeV);