@lukasmerten 2023-03-15
So basically with this PR simply the whole error calculation part is skipped as it would not be used anyway, right?