@reichherzerp 2023-03-15
Yes exactly, currently, we compute the propagation step (h) and the two small _helper_ propagation steps (h/2) to get the error estimate that we do not need when using fixed steps. By removing the two _helper_ steps, we save 2/3 of these time-consuming computations.