@lukasmerten 2023-03-16
I can roughly confirm the speed improvement. Using the above script with the SIMD optimization for the plane wave approach ("Fast waves"), decreases the performance a bit, but it still is a significant improvement.